style: allow type checking with TypeScript
TypeScript's artifacts are not used.
