+++ .prettierrc.yml
@@ -1,3 +1,5 @@
+plugins:
+  - prettier-plugin-jsdoc
 singleQuote: true
 bracketSpacing: true
 endOfLine: lf
