+++ package.json
@@ -120,6 +120,11 @@
   ],
   "devDependencies": {
     "@swellaby/eslint-config": "^2.0.0",
+    "@types/chai": "^4.3.5",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^18",
+    "@types/sinon": "^10.0.16",
+    "@types/vscode": "^1.49.0",
     "c8": "^8.0.1",
     "chai": "^4.3.7",
     "eslint": "^8.46.0",
@@ -127,8 +132,10 @@
     "mocha-multi-reporters": "^1.5.1",
     "mocha-sonarqube-reporter": "^1.0.2",
     "prettier": "^3.0.1",
+    "prettier-plugin-jsdoc": "^1.0.1",
     "rimraf": "^5.0.1",
     "sinon": "^15.2.0",
+    "typescript": "^5.1.6",
     "vsce": "^2.15.0"
   },
   "dependencies": {
@@ -136,7 +143,8 @@
     "jsonc-parser": "^3.2.0"
   },
   "scripts": {
-    "build": "npm test",
+    "build": "npm run test",
+    "build:tsc": "tsc --build",
     "clean": "rimraf .testresults .coverage .vsix",
     "coverage": "npm run coverage:unit",
     "coverage:unit": "c8 -c tests/unit/.c8rc.json npm run test:unit",
