+++ tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "files": [],
+  "include": [],
+  "references": [{ "path": "./src" }, { "path": "./tests/unit" }]
+}
